+++ Gruntfile.js
@@ -3,7 +3,6 @@
 /* eslint node/no-sync: 0 */
 require('./processRequire.js');
 const fs = require('fs');
-const openVeoApi = require('@openveo/api');
 
 /**
  * Loads a bunch of grunt configuration files from the given directory.
@@ -33,64 +32,13 @@ module.exports = function(grunt) {
     env: process.env
   };
 
-  // Set "withSourceMaps" property which will be used by grunt tasks to set appropriate configuration
-  process.withSourceMaps = (process.argv.length > 3 && process.argv[3] === '--with-source-maps') ? true : false;
-
   grunt.initConfig(config);
 
   // Load tasks definitions
   grunt.config.merge(loadConfig('./tasks'));
 
   // Load grunt tasks
-  grunt.loadNpmTasks('grunt-contrib-compass');
-  grunt.loadNpmTasks('grunt-contrib-watch');
-  grunt.loadNpmTasks('grunt-contrib-uglify');
-  grunt.loadNpmTasks('grunt-contrib-concat');
   grunt.loadNpmTasks('grunt-gh-pages');
   grunt.loadNpmTasks('grunt-protractor-runner');
-  grunt.loadNpmTasks('grunt-contrib-copy');
-  grunt.loadNpmTasks('grunt-replace');
-  grunt.loadNpmTasks('grunt-angular-templates');
-
-  grunt.registerMultiTask('remove', openVeoApi.grunt.removeTask(grunt));
-  grunt.registerMultiTask('ngDp', openVeoApi.grunt.ngDpTask(grunt));
-
-  // Dynamically set src property of the concat:back-office task
-  // The list of sources is built dynamically by the ngDp:back-office task
-  grunt.registerTask('back-office-set-concat-src', 'Set src of concat:back-office task', () => {
-
-    // Get the list of sources to concat from the results of ngDp:back-office task
-    const resources = process.require('build/ng-admin-files.json');
-    var concat = grunt.config('concat');
-    concat['back-office'].src = [];
-
-    resources.js.forEach((resourcePath) => {
-      concat['back-office'].src.push(`<%= project.adminSourcesPath %>/${resourcePath}`);
-    });
-
-    grunt.config('concat', concat);
-  });
-
-  // Build the front office
-  grunt.registerTask('build-front-office-client', [
-    'compass:front-office',
-    'uglify:front-office',
-    'concat:front-office-lib',
-    'concat:front-office-js'
-  ]);
-
-  // Build the back office
-  grunt.registerTask('build-back-office-client', [
-    'ngDp:back-office',
-    'replace:back-office-inject-scss',
-    'replace:back-office-inject-scripts',
-    'compass:back-office',
-    'replace:back-office-font-paths',
-    'back-office-set-concat-src',
-    'concat:back-office',
-    'uglify:back-office',
-    'copy:back-office-html-root',
-    'ngtemplates:back-office'
-  ]);
 
 };