+++ CHANGELOG.md
@@ -17,10 +17,17 @@
 - **dashjs** has been upgraded from 3.0.3 to **4.1.0**
 - **eslint** has been upgraded from 22.0.0 to **23.0.0**
 - **express-session** has been upgraded from 1.17.0 to **1.17.2**
+- **grunt-angular-templates** has been removed
+- **grunt-contrib-compass** has been removed
+- **grunt-contrib-concat** has been removed
+- **grunt-contrib-copy** has been removed
+- **grunt-contrib-uglify** has been removed
+- **grunt-contrib-watch** has been removed
 - **grunt-eslint** has been removed
 - **grunt-karma** has been removed
 - **grunt-mkdocs** has been removed
 - **grunt-mocha-test** has been removed
+- **grunt-replace** has been removed
 - **jsonpath-plus** has been upgraded from 3.0.0 to **6.0.1**
 - **karma** has been upgraded from 4.4.1 to **6.3.4**
 - **karma-mocha** has been upgraded from 1.3.0 to **2.0.1**